v.2.1.0 - The Cloud Update
We added 3 Softwares

2 Cloud Systems and 1 CMS.

* Updated WordPress
* Updated Typo3
* Updated Shopware
* Updated MyBB
* Updated EspoCRM
* Updated Joomla
+ Added OwnCloud
+ Added NextCloud
+ Added ImpressPages

+++ installer.php
@@ -1,83 +1,123 @@
 <?php 
-/**
- * CMS-Installer
- * Copyright (c) Webintosh.at | Julian Seidl
- *
- * Version 2.0.0
- *
- * According to our dual licensing model, this program can be used either
- * under the terms of the GNU Affero General Public License, version 3,
- * or under a proprietary license.
- *
- * The texts of the GNU Affero General Public License with an additional
- * permission and of our proprietary license can be found at and
- * in the LICENSE file you have received along with this program.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- */
+/*
+*
+* CMS-Installer
+* Copyright (c) jseidl.at | Julian Seidl
+*
+* MIT License
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+* 
+* The above copyright notice and this permission notice shall be included in all
+* copies or substantial portions of the Software.
+* 
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*
+* Version 2.1.0
+*
+*/
 	// Database
-	$version = "v2.0.0";
+	$version = "v2.1.0";
 	$apps = array(
 		"WordPress" => array(
 			"name" => "WordPress",
-			"version" => "4.7.7",
-			"category" => "Blog",
+			"version" => "4.7.6",
+			"category" => "Blog / CMS",
 			"link" => "https://wordpress.org",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/wordpress.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/wordpress-4.7.4.tar.gz"
+			"img" => "https://cms-installer.webintosh.at/img/wordpress.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/wordpress-4.7.6.tar.gz"
 		),
 		"Typo3-7" => array(
 			"name" => "Typo3",
-			"version" => "7.6.16",
+			"version" => "7.6.18",
 			"category" => "CMS",
 			"link" => "https://typo3.org",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/typo3.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/t3-7.6.16.tar.gz"
+			"img" => "https://cms-installer.webintosh.at/img/typo3.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/t3-7.6.18.tar.gz"
 		),
 		"MediaWiki" => array(
 			"name" => "MediaWiki",
 			"version" => "1.28.2",
 			"description" => "",
 			"category" => "Wiki",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/mediawiki.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/mediawiki-1.28.2.tar.gz"
+			"img" => "https://cms-installer.webintosh.at/img/mediawiki.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/mediawiki-1.28.2.tar.gz"
 		),
 		"Shopware" => array(
 			"name" => "Shopware",
-			"version" => "5.22.2",
+			"version" => "5.2.24",
 			"category" => "E-Commerce",
 			"link" => "https://de.shopware.com",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/shopware.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/sw-5.22.2.zip"
+			"img" => "https://cms-installer.webintosh.at/img/shopware.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/sw-5.2.24.zip"
 		),
 		"Espocrm" => array(
 			"name" => "Espocrm",
-			"version" => "4.6.0",
+			"version" => "4.7.0",
 			"category" => "CRM",
 			"link" => "https://www.espocrm.com/",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/espocrm.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/EspoCRM-4.6.0.zip"
+			"img" => "https://cms-installer.webintosh.at/img/espocrm.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/EspoCRM-4.7.0.zip"
 		),
 		"Joomla" => array(
 			"name" => "Joomla!",
-			"version" => "3.7.0",
+			"version" => "3.7.2",
 			"category" => "CMS",
 			"link" => "https://www.joomla.de/",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/joomla.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/joomla-3.7.0.zip"
+			"img" => "https://cms-installer.webintosh.at/img/joomla.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/joomla-3.7.2.zip"
 		),
 		"Mybb" => array(
 			"name" => "MyBB",
-			"version" => "1.8.11",
+			"version" => "1.8.12",
 			"category" => "Forum",
 			"link" => "https://mybb.com",
-			"img" => "https://cms-installer.webintosh.at/2.0.0/img/mybb.png",
-			"download" => "https://cms-installer.webintosh.at/2.0.0/mybb-1.8.11.zip"
+			"img" => "https://cms-installer.webintosh.at/img/mybb.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/mybb-1.8.12.zip"
 		),
+		"Impresspages" => array(
+			"name" => "ImpressPages",
+			"version" => "5.2.0",
+			"category" => "CMS",
+			"link" => "https://www.impresspages.org",
+			"img" => "https://cms-installer.webintosh.at/img/impresspages.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/ip-5.2.0.zip"
+		),
+		"NextCloud" => array(
+			"name" => "MyBB",
+			"version" => "12.0.0",
+			"category" => "Cloud / Datastorage",
+			"link" => "https://nextcloud.com",
+			"img" => "https://cms-installer.webintosh.at/img/nextcloud.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/nextcloud-12.0.0.zip"
+		),
+		"OwnCloud" => array(
+			"name" => "OwnCloud",
+			"version" => "10.0.1",
+			"category" => "Cloud / Datastorage",
+			"link" => "https://owncloud.org",
+			"img" => "https://cms-installer.webintosh.at/img/owncloud.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/owncloud-10.0.1.zip"
+		),
+		"Grav" => array(
+			"name" => "Grav",
+			"version" => "1.2.4",
+			"category" => "CMS",
+			"link" => "https://owncloud.org",
+			"img" => "https://cms-installer.webintosh.at/img/grav.png",
+			"download" => "https://cms-installer.webintosh.at/2.1.0/grav-1.2.4.zip"
+		)
 	);
 	$httpMethod = filter_input(INPUT_SERVER, 'REQUEST_METHOD');
     $cms = filter_input(INPUT_POST, 'cms');
@@ -162,17 +202,17 @@
     			}
     		}
     	}
-    	else if("Joomla" === $cms){
-    		foreach ($apps as $items => $item) {
-    			if($items === $cms){
-    				$pathname = dirname($_SERVER['PHP_SELF']);
-    				exec('wget '. $item["download"] .' ');
-    				exec('unzip joomla-3.7.0.zip');
-                    exec('mv Joomla_3.7.0-Stable-Full_Package/* .');
-                    exec('rm -rf Joomla_3.7.0-Stable-Full_Package/');
-                    exec('rm -rf joomla-3.7.0.zip');
-                    exec('chmod -R 775 ..'. $pathname .'');
-                    exec('rm -rf installer.php');
+		else if("Joomla" === $cms){
+			foreach ($apps as $items => $item) {
+				if($items === $cms){
+					$pathname = dirname($_SERVER['PHP_SELF']);
+					exec('wget '. $item["download"] .' ');
+					exec('unzip joomla-3.7.0.zip');
+					exec('mv Joomla_3.7.0-Stable-Full_Package/* .');
+					exec('rm -rf Joomla_3.7.0-Stable-Full_Package/');
+					exec('rm -rf joomla-3.7.0.zip');
+					exec('chmod -R 775 ..'. $pathname .'');
+					exec('rm -rf installer.php');
     			}
     		}
     	}
@@ -185,6 +225,62 @@
                     exec('mv mybb-1.8.11/* .');
                     exec('rm -rf mybb-1.8.11/');
                     exec('rm -rf mybb-1.8.11.zip');
+                    exec('chmod -R 775 ..'. $pathname .'');
+                    exec('rm -rf installer.php');
+    			}
+    		}
+    	}
+    	else if("Impresspages" === $cms){
+    		foreach ($apps as $items => $item) {
+    			if($items === $cms){
+    				$pathname = dirname($_SERVER['PHP_SELF']);
+    				exec('wget '. $item["download"] .' ');
+    				exec('unzip ip-4.10.1.zip');
+                    exec('mv ImpressPages/* .');
+                    exec('rm -rf ImpressPages/');
+                    exec('rm -rf ip-4.10.1.zip');
+                    exec('chmod -R 775 ..'. $pathname .'');
+                    exec('rm -rf installer.php');
+    			}
+    		}
+    	}
+    	else if("NextCloud" === $cms){
+    		foreach ($apps as $items => $item) {
+    			if($items === $cms){
+    				$pathname = dirname($_SERVER['PHP_SELF']);
+    				exec('wget '. $item["download"] .' ');
+    				exec('unzip nextcloud-11.0.3.zip');
+                    exec('mv nextcloud/* .');
+                    exec('rm -rf nextcloud/');
+                    exec('rm -rf nextcloud-11.0.3.zip');
+                    exec('chmod -R 775 ..'. $pathname .'');
+                    exec('rm -rf installer.php');
+    			}
+    		}
+    	}
+    	else if("OwnCloud" === $cms){
+    		foreach ($apps as $items => $item) {
+    			if($items === $cms){
+    				$pathname = dirname($_SERVER['PHP_SELF']);
+    				exec('wget '. $item["download"] .' ');
+    				exec('unzip owncloud-10.0.0.zip');
+                    exec('mv owncloud/* .');
+                    exec('rm -rf owncloud/');
+                    exec('rm -rf owncloud-10.0.0.zip');
+                    exec('chmod -R 775 ..'. $pathname .'');
+                    exec('rm -rf installer.php');
+    			}
+    		}
+    	}
+    	else if("Grav" === $cms){
+    		foreach ($apps as $items => $item) {
+    			if($items === $cms){
+    				$pathname = dirname($_SERVER['PHP_SELF']);
+    				exec('wget '. $item["download"] .' ');
+    				exec('unzip grav-1.2.4.zip');
+                    exec('mv grav-admin/* .');
+                    exec('rm -rf grav-admin/');
+                    exec('rm -rf grav-1.2.4.zip');
                     exec('chmod -R 775 ..'. $pathname .'');
                     exec('rm -rf installer.php');
     			}
@@ -269,13 +365,13 @@
 			<div class="container" style="height: 80%; min-height: 600px;">
 				<div class="row" style="height: 80%; min-height: 600px;">
 					<div class="col-md-12" style="height: 80%; min-height: 600px;">
-						<div class="display-table">
+						<div class="display-table" style="height: 80%; min-height: 600px;">
 							<div class="display-table-cell">
 								<div class="progress">
 									<div class="progress-bar progress-bar-info progress-bar-striped <?php echo $installed; ?>" role="progressbar" aria-valuenow="20" aria-valuemin="0" aria-valuemax="100"></div>
 								</div>
 								<h2><i class="fa fa-check-square-o" aria-hidden="true"></i> Installed</h2>
-								<p><a href="/">Klick me</a>to go to your new CMS</p>
+								<p><a href="/">Click me</a> to go to your new CMS</p>
 							</div>
 						</div>
 					</div>
@@ -286,14 +382,14 @@
 			<div class="container-fluid">
 				<div class="row">
 					<div class="col-md-4">
-						<p>&copy; <?php echo date("Y"); ?> by <a href="https://jseidl.at">Julian Seidl</a></p>
+						<p>&copy; <?php echo date("Y"); ?> by <a href="https://jseidl.at">jSeidl</a></p>
 					</div>
 					<div class="col-md-4 text-center">
 						<p><?php echo($version); ?></p>
 					</div>
 					<div class="col-md-4 text-right">
 						<ul>
-							<li><a href="https://github.com/Thejuse/CMS-Installer/issues">Bug report</a></li>
+							<li><a href="https://github.com/webintosh/CMS-Installer/issues">Bug report</a></li>
 						</ul>
 					</div>
 				</div>
